requirements.txt
asn1crypto==0.24.0
boxsdk==1.5.5
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
cryptography==2.1.4
idna==2.6
lxml==4.1.1
pycparser==2.18
PyJWT==1.5.3
PyPdf2==1.26
python-dateutil==2.6.1
python-docx==0.8.6
requests==2.18.4
requests-toolbelt==0.8.0
six==1.11.0
urllib3==1.22
watson-developer-cloud==2.5.4
